Make use of Sopel 7 tools.time.seconds_to_human() in cooldown
Not quite sure I like this, since it force-inserts a preposition instead
of letting the caller handle formatting the result. Need to think about
the possibility of changing Sopel's API (adding a "cleaner" function to
get just the human-readable units) or just stealing my AniTV plugin's
code to do this instead.