Migrate from `bot.privileges` to `bot.channels`
Sopel 8.0 removed `bot.privileges` entirely, so the plugin stopped
working when users updated their bots.

There's still a good bit of modernization to do here, and Sopel 7.1
support is likely not long for this world since 8.0 added the helper
method `bot.make_identifier()`; this is just the minimum tweaking needed
to make the plugin work on up-to-date bots.