Refactor a bunch of helpers into a manager class + Sopel 8 only
Went Sopel 8+ only because of the new `bot.make_identifier()` helper
method, and the stronger checks around `SopelIdentifierMemory` keys
(used for things like channels/users), and the better enforcement in
`bot.db` of making sure the nick passed to `*_nick_value` methods is
correctly treated as an `Identifier`.

Basically, Sopel 7 is still a bit of a slopfest when it comes to dealing
with `Identifier` values correctly when compared to 8. I wanted to get
rid of as many typecasts in the plugin code as possible.